Widen raw_events.bytes to BIGINT so a big download can't wedge the aggregator
A single large transfer (an OS/browser update, a video) logs a %<st byte
count past the int32 ceiling (~2.1B), and a long-lived CONNECT tunnel's %tr
can too. On Postgres that overflowed the INTEGER `bytes` column and asyncpg
raised DataError for the whole executemany() batch. flush() never advances
_last_flushed_id past an uncommitted window, so the aggregator re-ran the
same poisoned 2000-row batch every interval forever -- the unflushed
backlog climbed until the ring buffer overflowed and events were dropped
without ever being persisted (the reported "backlog_ratio 0.9 -> 1.0+ ->
520k events lost"). Every aggregate table's total_bytes was already
BigInteger; raw_events.bytes and .duration_ms were the ones missed.

Three layers:
- Migration + model: raw_events.bytes and .duration_ms -> BIGINT (Postgres;
  SQLite dev/test builds from the models and its INTEGER is already 64-bit).
- Parser: clamp the numeric fields to [0, 2**63-1], the same "sanitise so
  the INSERT can't fail" contract the string fields already follow via
  _truncate -- a corrupt/absurd token can't become an un-persistable row.
- Aggregator: a DataError on a raw_events batch is now caught, the batch
  retried row-by-row in savepoints, and only the individual failing rows
  dropped (loudly) instead of the whole flush aborting. Any other error
  still propagates as a retry -- a transient fault must not silently drop.

+++ backend/app/models/raw_event.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Boolean, Index, Integer, String, Text
+from sqlalchemy import BigInteger, Boolean, Index, Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.config import DEFAULT_BRANCH
@@ -35,12 +35,21 @@ class RawEvent(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     timestamp: Mapped[datetime] = mapped_column(UTCDateTime, index=True)
-    duration_ms: Mapped[int] = mapped_column(Integer)
+    # BigInteger, not Integer: a single large file download (a Windows/Chrome
+    # update, a video) logs a %<st byte count well past the ~2.1B int32
+    # ceiling, and a long-lived CONNECT tunnel's %tr can too. An int32
+    # column turned one such row into an asyncpg DataError that failed the
+    # whole flush batch, and since the aggregator never advances past an
+    # uncommitted flush it retried the same poisoned batch every interval
+    # forever -- backlog growing until the ring buffer overflowed and
+    # events were lost for good. Every aggregate table's total_bytes was
+    # already BigInteger; this column was the one that got missed.
+    duration_ms: Mapped[int] = mapped_column(BigInteger)
     client_ip: Mapped[str] = mapped_column(String(45), index=True)
     branch: Mapped[str] = mapped_column(String(64), default=DEFAULT_BRANCH)
     action: Mapped[str] = mapped_column(String(64))
     status_code: Mapped[int] = mapped_column(Integer)
-    bytes: Mapped[int] = mapped_column(Integer)
+    bytes: Mapped[int] = mapped_column(BigInteger)
     method: Mapped[str] = mapped_column(String(16))
     url: Mapped[str] = mapped_column(Text)
     domain: Mapped[str | None] = mapped_column(String(255), nullable=True, index=True)
